server/requirements.txt
fastapi
uvicorn
transformers
Pillow
requests
gradio
torch 
